requirements.txt
python-telegram-bot==21.3
playwright==1.41.0
python-dotenv==1.0.1
